// app/types/type.ts
export type Type = {
  slot: number;
  type: {
    name:
      | 'normal'
      | 'fire'
      | 'water'
      | 'grass'
      | 'electric'
      | 'ice'
      | 'fighting'
      | 'poison'
      | 'ground'
      | 'flying'
      | 'psychic'
      | 'bug'
      | 'rock'
      | 'ghost'
      | 'dark'
      | 'dragon'
      | 'steel'
      | 'fairy';
    url: string;
  };
};